SIMD refactor
